Turn App into React.Component
