+++ packages/ui-app/src/index.tsx
@@ -2,12 +2,6 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import '@polkadot/ui-settings';
-import './i18n';
-import './styles';
-
-import { BareProps } from './types';
-
 import React, { Suspense } from 'react';
 import ReactDOM from 'react-dom';
 import { HashRouter } from 'react-router-dom';
@@ -48,39 +42,3 @@ export { default as Progress } from './Progress';
 export { default as Static } from './Static';
 export { default as Status } from './Status';
 export { default as Tabs } from './Tabs';
-
-type Props = BareProps & {
-  url?: string
-};
-
-export default function createApp (App: React.ComponentType<BareProps>, { className, style, url }: Props, rootId: string = 'root'): void {
-  const rootElement = document.getElementById(rootId);
-
-  if (!rootElement) {
-    throw new Error(`Unable to find element with id '${rootId}'`);
-  }
-
-  ReactDOM.render(
-    <Suspense fallback='...'>
-      <Queue>
-        <QueueConsumer>
-          {({ queueExtrinsic, queueSetTxStatus }) => (
-            <Api
-              queueExtrinsic={queueExtrinsic}
-              queueSetTxStatus={queueSetTxStatus}
-              url={url}
-            >
-              <HashRouter>
-                <App
-                  className={className}
-                  style={style}
-                />
-              </HashRouter>
-            </Api>
-          )}
-        </QueueConsumer>
-      </Queue>
-    </Suspense>,
-    rootElement
-  );
-}
